# Adding interpreter callback