--- a/src/bin/brainfuck.rs
+++ b/src/bin/brainfuck.rs
@@ -9,10 +9,12 @@ use std::path::{Path};
 use std::fs::File;
 use std::io;
 use std::io::prelude::*;
+use std::thread;
+use std::time::Duration;
 
 use clap::{Arg, App};
 
-use brainfuck::{precompile, Interpreter};
+use brainfuck::{precompile, interpret};
 
 macro_rules! exit_with_error(
     ($($arg:tt)*) => { {
@@ -80,11 +82,23 @@ fn main() {
     let mut bytes = Vec::new();
     f.read_to_end(&mut bytes).expect("Fatal: Could not read source file");
     let program = precompile(bytes.iter(), opt);
-    Interpreter::new(
+    interpret(
         io::stdin(),
         io::stdout(),
-        io::stderr(),
-        debug_mode,
-        delay,
-    ).interpret(program);
+        program,
+        |state| {
+            if debug_mode {
+                writeln!(
+                    &mut io::stderr(),
+                    "{{\"nextInstructionIndex\": {}, \"lastInstruction\": {}, \"currentPointer\": {}, \"memory\": \"{}\"}}",
+                    state.next_instruction,
+                    state.last_instruction.map_or_else(|| "null".to_owned(), |instr| format!("\"{}\"", instr.to_string())),
+                    state.current_pointer,
+                    state.memory.iter().fold(String::new(), |acc, v| format!("{} {}", acc, v))
+                ).expect("failed printing to stderr");
+            }
+
+            thread::sleep(Duration::from_millis(delay));
+        }
+    );
 }