--- a/src/bin/brainfuck.rs
+++ b/src/bin/brainfuck.rs
@@ -2,6 +2,7 @@
 
 #[macro_use]
 extern crate clap;
+extern crate colored;
 
 extern crate brainfuck;
 
@@ -12,9 +13,10 @@ use std::io::prelude::*;
 use std::thread;
 use std::time::Duration;
 
+use colored::*;
 use clap::{Arg, App};
 
-use brainfuck::{precompile, interpret};
+use brainfuck::{precompile, interpret, InterpreterState, DebugFormat, Instruction, OptimizationLevel};
 
 macro_rules! exit_with_error(
     ($($arg:tt)*) => { {
@@ -42,6 +44,13 @@ fn main() {
             .long("debug")
             .help("Enables debug mode which outputs debugging information to stderr")
         )
+        .arg(Arg::with_name("debug-format")
+            .long("debug-format")
+            .value_name("format")
+            .default_value("text")
+            .possible_values(&["text", "json"])
+            .help("The format of the debugging output")
+        )
         .arg(Arg::with_name("optimize")
             .short("O")
             .long("optimize")
@@ -65,6 +74,8 @@ fn main() {
     }
 
     let debug_mode = args.is_present("debug-enabled");
+    // We can call unwrap() because the validation is already done by clap
+    let debug_format = args.value_of("debug-format").unwrap().parse().unwrap();
 
     let delay: u64 = if let Some(delay_str) = args.value_of("delay") {
         delay_str.parse().unwrap_or_else(|e: std::num::ParseIntError| exit_with_error!("Invalid delay: {}", e))
@@ -83,29 +94,22 @@ fn main() {
     f.read_to_end(&mut bytes).expect("Fatal: Could not read source file");
     let program = precompile(bytes.iter(), opt);
 
-    // Based on debug_mode and delay, this will run one of three functions
+    // Based on debug_mode and delay, this will run one of several functions
     // If there is no delay and debug mode is off, performance is prioritized and the interpreter
     // should run at top speed
     let input = io::stdin();
     let output = io::stdout();
     if debug_mode {
-        interpret(
-            input,
-            output,
-            program,
-            |state| {
-                writeln!(
-                    &mut io::stderr(),
-                    "{{\"nextInstructionIndex\": {}, \"lastInstruction\": {}, \"currentPointer\": {}, \"memory\": \"{}\"}}",
-                    state.next_instruction,
-                    state.last_instruction.map_or_else(|| "null".to_owned(), |instr| format!("\"{}\"", instr.to_string())),
-                    state.current_pointer,
-                    state.memory.iter().fold(String::new(), |acc, v| format!("{} {}", acc, v))
-                ).expect("failed printing to stderr");
+        match debug_format {
+            DebugFormat::Text => {
+                interpret(input, output, program, |state| format_human_readable(state, delay, match opt {
+                    OptimizationLevel::Off => 1,
+                    OptimizationLevel::Speed => 4,
+                }));
+            },
 
-                thread::sleep(Duration::from_millis(delay));
-            }
-        );
+            DebugFormat::Json => interpret(input, output, program, |state| format_json(state, delay)),
+        }
     }
     // Need this condition because delay can be active without debug_mode
     else if delay > 0 {
@@ -114,4 +118,58 @@ fn main() {
     else {
         interpret(input, output, program, |_| {});
     }
+}
+
+#[inline]
+fn format_human_readable(state: InterpreterState, delay: u64, instruction_width: usize) {
+    use Instruction::*;
+
+    let pointer = state.current_pointer;
+
+    let current_instruction = format!("#{:<3}", state.current_instruction);
+
+    let instr = state.instruction;
+    let instruction = match instr {
+        Right(..) | Left(..) => instr.to_string().on_cyan(),
+        Increment(..) | Decrement(..) => instr.to_string().on_green(),
+        Write => instr.to_string().on_purple(),
+        Read => instr.to_string().on_yellow(),
+        JumpForwardIfZero { .. } | JumpBackwardUnlessZero { .. } => {
+            instr.to_string().on_blue()
+        },
+    }.bold();
+
+    let memory = state.memory.iter().enumerate().fold(String::new(), |acc, (i, c)| {
+        let mut cell = c.to_string().normal();
+        if i == pointer {
+            cell = cell.blue().bold();
+        }
+        format!("{} {:>3}", acc, cell)
+    });
+
+    writeln!(
+        &mut io::stderr(),
+        "{} {:instruction_width$} {}",
+        current_instruction.normal(),
+        instruction,
+        memory,
+
+        instruction_width = instruction_width,
+    ).expect("failed to write debug output to stderr");
+
+    thread::sleep(Duration::from_millis(delay));
+}
+
+#[inline]
+fn format_json(state: InterpreterState, delay: u64) {
+    writeln!(
+        &mut io::stderr(),
+        "{{\"currentInstructionIndex\": {}, \"instruction\": \"{}\", \"currentPointer\": {}, \"memory\": \"{}\"}}",
+        state.current_instruction,
+        state.instruction,
+        state.current_pointer,
+        state.memory.iter().fold(String::new(), |acc, v| format!("{} {}", acc, v))
+    ).expect("failed printing to stderr");
+
+    thread::sleep(Duration::from_millis(delay));
 }