--- a/src/instruction.rs
+++ b/src/instruction.rs
@@ -1,5 +1,4 @@
 use std::fmt;
-use std::iter::repeat;
 
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum Instruction {
@@ -31,17 +30,22 @@ pub enum Instruction {
 
 impl fmt::Display for Instruction {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", match *self {
-            Instruction::Right(n) => repeat(">").take(n).collect(),
-            Instruction::Left(n) => repeat("<").take(n).collect(),
-            Instruction::Increment(n) => repeat("+").take(n).collect(),
-            Instruction::Decrement(n) => repeat("-").take(n).collect(),
+        f.write_str(match *self {
+            Instruction::Right(n) => format_instruction(">", n),
+            Instruction::Left(n) => format_instruction("<", n),
+            Instruction::Increment(n) => format_instruction("+", n),
+            Instruction::Decrement(n) => format_instruction("-", n),
             Instruction::Write => ".".to_owned(),
             Instruction::Read => ",".to_owned(),
             Instruction::JumpForwardIfZero { .. } => "[".to_owned(),
             Instruction::JumpBackwardUnlessZero { .. } => "]".to_owned(),
-        })
+        }.as_ref())
     }
+}
+
+#[inline]
+fn format_instruction(instr: &str, n: usize) -> String {
+    format!("{}{}", instr, if n > 1 {format!("{}", n)} else {"".to_owned()})
 }
 
 #[cfg(test)]
@@ -51,20 +55,20 @@ mod tests {
     #[test]
     fn display() {
         assert_eq!(Instruction::Right(1).to_string(), ">");
-        assert_eq!(Instruction::Right(2).to_string(), ">>");
-        assert_eq!(Instruction::Right(5).to_string(), ">>>>>");
+        assert_eq!(Instruction::Right(2).to_string(), ">2");
+        assert_eq!(Instruction::Right(5).to_string(), ">5");
 
         assert_eq!(Instruction::Left(1).to_string(), "<");
-        assert_eq!(Instruction::Left(2).to_string(), "<<");
-        assert_eq!(Instruction::Left(5).to_string(), "<<<<<");
+        assert_eq!(Instruction::Left(2).to_string(), "<2");
+        assert_eq!(Instruction::Left(5).to_string(), "<5");
 
         assert_eq!(Instruction::Increment(1).to_string(), "+");
-        assert_eq!(Instruction::Increment(2).to_string(), "++");
-        assert_eq!(Instruction::Increment(5).to_string(), "+++++");
+        assert_eq!(Instruction::Increment(2).to_string(), "+2");
+        assert_eq!(Instruction::Increment(5).to_string(), "+5");
 
         assert_eq!(Instruction::Decrement(1).to_string(), "-");
-        assert_eq!(Instruction::Decrement(2).to_string(), "--");
-        assert_eq!(Instruction::Decrement(5).to_string(), "-----");
+        assert_eq!(Instruction::Decrement(2).to_string(), "-2");
+        assert_eq!(Instruction::Decrement(5).to_string(), "-5");
 
         assert_eq!(Instruction::Write.to_string(), ".");
         assert_eq!(Instruction::Read.to_string(), ",");