--- a/src/optlevel.rs
+++ b/src/optlevel.rs
@@ -1,6 +1,6 @@
 use std::str::FromStr;
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum OptimizationLevel {
     // Do not optimize
     Off,