Fixes #6

This PR adds a callback to the interpreter and refactors it back into a simple function. The entire struct with one method was completely unnecessary and the iterator approach in #15 slowed everything down a lot.

Hopefully this approach fixes things so everything is faster and still elegant.

This is an important change because it separates the debug and delay logic from the interpreter and makes everything testable.

----

**Discovered and fixed a major bug in the debugging output.**

- [x] Check profiling output and evaluate how to move forward based on speed difference
- [x] Update [brain-lang/brain-debug](https://github.com/brain-lang/brain-debug) to accept the new format and add something to the README about a minimum version (see https://github.com/brain-lang/brain-debug/commit/df8de307f9f8fe3a50ccd29ebca6f5cbba7e63c4)
- [x] Make sure this bug gets fixed even if this PR is closed