@sunjay 2017-04-27
Timing results for #15, #16 (this PR) and master:

| time | master    | debug-output | debug-output2 |
| ---- | --------- | ------------ | ------------- |
| real | 0m47.054s | 0m59.973s    | 1m8.958s      |
| user | 0m46.404s | 0m59.392s    | 1m8.632s      |
| sys  | 0m0.640s  | 0m0.560s     | 0m0.288s      |