@sunjay 2017-04-27
All the trivial benchmarks sped up quite a bit:

```
$ cargo benchcmp master debug-output2 
 name                       master ns/iter  debug-output2 ns/iter  diff ns/iter   diff % 
 b01_compile_trivial        2,806           2,788                           -18   -0.64% 
 b01_compile_trivial_opt    1,682           1,656                           -26   -1.55% 
 b02_compile_large          2,087,058       2,080,122                    -6,936   -0.33% 
 b02_compile_large_opt      804,306         813,089                       8,783    1.09% 
 b03_compile_huge           67,373,079      66,493,848                 -879,231   -1.31% 
 b03_compile_huge_opt       14,001,127      13,989,128                  -11,999   -0.09% 
 b04_compile_simple         43,528          43,672                          144    0.33% 
 b04_compile_simple_opt     22,172          22,552                          380    1.71% 
 b05_compile_slow           269,560         269,369                        -191   -0.07% 
 b05_compile_slow_opt       135,005         134,974                         -31   -0.02% 
 b06_interpret_trivial      11,858          6,861                        -4,997  -42.14% 
 b06_interpret_trivial_opt  6,188           3,623                        -2,565  -41.45% 
 b07_interpret_simple       7,462,466       4,642,208                -2,820,258  -37.79% 
 b07_interpret_simple_opt   3,818,076       2,341,104                -1,476,972  -38.68%
```

![screenshot from 2017-04-26 22-01-39](https://cloud.githubusercontent.com/assets/530939/25464739/f3c1d6bc-2acb-11e7-9fe6-b2aa561174d9.png)