@sunjay 2017-04-27
Thanks to https://github.com/brain-lang/brainfuck/pull/16/commits/96088157a37830112c7827e2bebdeaa39b268481 there was a huge speed up when running `examples/mandel.bf`. The program now runs in 15.576 seconds. That's a 66.43% increase over master.