@sunjay 2017-04-27
The [benchmark results](https://travis-ci.org/brain-lang/brainfuck/jobs/226462199#L1119) look good too. Solid improvement in all the interpreter benchmarks.

![screenshot from 2017-04-27 15-08-45](https://cloud.githubusercontent.com/assets/530939/25500023/6eb8fae6-2b5b-11e7-86e0-a9abf9962e73.png)

```
 name                       previous-benchmark ns/iter  current-benchmark ns/iter  diff ns/iter   diff % 
 b01_compile_trivial        2,643                       2,600                               -43   -1.63% 
 b01_compile_trivial_opt    1,578                       1,531                               -47   -2.98% 
 b02_compile_large          2,078,859                   2,060,156                       -18,703   -0.90% 
 b02_compile_large_opt      807,003                     818,309                          11,306    1.40% 
 b03_compile_huge           82,403,200                  81,084,936                   -1,318,264   -1.60% 
 b03_compile_huge_opt       13,071,680                  13,324,260                      252,580    1.93% 
 b04_compile_simple         40,611                      40,660                               49    0.12% 
 b04_compile_simple_opt     21,473                      21,209                             -264   -1.23% 
 b05_compile_slow           256,119                     259,141                           3,022    1.18% 
 b05_compile_slow_opt       128,100                     128,466                             366    0.29% 
 b06_interpret_trivial      11,409                      6,848                            -4,561  -39.98% 
 b06_interpret_trivial_opt  5,772                       3,529                            -2,243  -38.86% 
 b07_interpret_simple       7,129,947                   4,466,058                    -2,663,889  -37.36% 
 b07_interpret_simple_opt   3,569,596                   2,216,498                    -1,353,098  -37.91% 
```